+++ changelog.md
@@ -25,6 +25,7 @@ Codion Change Log
 - EntityTableExportPanel configuration now included in user preferences.
 - EntityTableExportPanel, export to file added.
 - EntityTableExportPanel, alt-click toggles attribute inclusion, help text added, minor improvements.
+- EntityTableExportPanel, success messages added, minor fixes.
 
 ## 0.18.51
 ### is.codion.swing.common.ui